Various UI fixes
Content was wrongly set in a bold font; now it's set in a regular font.
Textual display of page views,visits and bounces looked bad; now it's properly aligned.
Added more visual clues to the interval header to make it obvious that the top pages and referrers are also from the specified interval.
Made label creation more versatile.
Presenters expose UI content always as ReactiveCocoa properties - no matter if it changes over time or not.